fix: adds a package stub to the _hygencook dir to better handle ts files
Hygen can parse a .ts prompt file in a template using ts-node/register, which is cool, but if you're
running templates in a directory where your package json has type: module that causes
ts-node/register to transpile to imports/exports instead of the commonjs format hygen needs. This
change drops a package.json stub into the _hygencook dir containing only type: "commonjs", which is
enough to convince ts-node/register to transpile any .ts files to the correct syntax for hygen to be
able to load.